read pipes before waiting for process output
So this was a really dumb bug where I didn't start reading from the pipes
until the subprocess had already exited. Which meant if the subprocess wrote
more bytes than was available in the kernel buffer -- 65536 bytes, at least on
my computer -- its writes would start to block, and it would never terminate.